♻️ strips the summary table to its own component
